requirements.txt
fastapi
uvicorn
autogenstudio
autogen-agentchat
autogen-ext
openai
pydantic
python-multipart
python-dotenv
cassandra-driver
aiohttp
pytest
pytest-asyncio
firebase-admin
python-jose[cryptography]
python-multipart
PyGithub
nanoid

